Handle `C.TIME_END_OF_SOURCE` buffer timestamps in `CeaDecoder`
The behaviour was changed in 1.4.0 with https://github.com/androidx/media/commit/bd9505b2cbd80121eccbf6975a3183f6d462db68,
so that the buffer timestamp is compared to `outputStartTimeUs` when
deciding whether to discard a "decode only" buffer before decoding
(instead of the deprecated/removed `isDecodeOnly` property). This breaks
when the buffer timestamp is `TIME_END_OF_SOURCE` (which is
`Long.MIN_VALUE`), because `TIME_END_OF_SOURCE < outputStartTimeUs` is
always true, so the end-of-stream buffer is never passed to the decoder
and on to `TextRenderer` where it is used to
[set `inputStreamEnded = true`](https://github.com/androidx/media/blob/6bb4008267af9f91224a6510cfc7e96bae459a03/libraries/exoplayer/src/main/java/androidx/media3/exoplayer/text/TextRenderer.java#L434-L436)
and so playback hangs.

Issue: androidx/media#1863
PiperOrigin-RevId: 695767247
(cherry picked from commit 51d02f110fa7de5f83810f30c6e3d3acac4f36cb)