Use live providers in CompositionPlayer to remove workaround
When CompositionPlayer added repeat modes in https://github.com/androidx/media/commit/d1af406b3d7bf0c0288ad51a41d7d072e5be3953, it
changed the logic in SimpleBasePlayer to workaround the specific
scenario in CompositionPlayer that didn't seem to work correctly.
The workaround is not really generically applicable though. The
source of the original problem was that the position values of
all state objects were connected to the "live" source of the
current player so that the repetition case couldn't be detected
(the position before and after the seek looked the same). The
solution is to use the newly added LivePositionSuppliers and
disconnect them before seeking.

PiperOrigin-RevId: 713659638